Update: Finish moving fili/elide metadata to client
