+++ packages/client/src/plugins/elide-apollo-client.ts
@@ -3,18 +3,23 @@
  * Licensed under the terms of the MIT license. See accompanying LICENSE.md file for terms.
  */
 
-import { ApolloClient, InMemoryCache, HttpLink } from '@apollo/client/core';
-import { setContext } from '@apollo/client/link/context';
+import { ApolloClient, InMemoryCache, HttpLink } from '@apollo/client/core/index.js';
+import { setContext } from '@apollo/client/link/context/index.js';
 import type { ClientConfig } from '../config/datasources.js';
+import fetch from 'cross-fetch';
 
 export default function getClient(config: ClientConfig) {
   function buildURLPath(dataSourceName: string): string {
     const host = config.configHost({ dataSourceName });
-    return new URL(host, `${document.location.protocol}${document.location.host}`).href;
+    let baseUrl;
+    if (document) {
+      baseUrl = `${document.location.protocol}${document.location.host}`;
+    }
+    return new URL(host, baseUrl).href;
   }
 
   function link() {
-    const linkOptions = { credentials: 'include' };
+    const linkOptions = { credentials: 'include', fetch };
     const httpLink = new HttpLink(linkOptions);
 
     const contextLink = setContext((_, context) => {
