+++ packages/data/addon/adapters/metadata/elide.ts
@@ -8,6 +8,7 @@ import ElideMetadataAdapter from '@yavin/client/plugins/elide/adapters/metadata'
 export default class NaviElideMetadataAdapter extends ElideMetadataAdapter {
   static create(args: unknown) {
     const owner = getOwner(args);
-    return new this(owner.lookup('service:client-injector'));
+    const yavinClient = owner.lookup('service:yavin-client');
+    return new this(yavinClient.injector);
   }
 }
